fix warning and bug in waypointReference module
Former-commit-id: 01b749156eaddea29655d667bb94e84d5411a7f9
Former-commit-id: 2d1295b5454678e22d8f480a4ec9c279175f1e3b

+++ src/fswAlgorithms/attGuidance/waypointReference/waypointReference.cpp
@@ -102,7 +102,7 @@ void WaypointReference::UpdateState(uint64_t CurrentSimNanos)
         attMsgBuffer = this->attRefOutMsg.zeroMsgPayload;
 		
 		/* current time */
-		double t = CurrentSimNanos;
+		double t = (double) (CurrentSimNanos);
 		
 		/* for CurrentTime < t_0 hold initial attitude with zero angular rates and accelerations */
 		if (t < this->t_a) {
